+++ src/App.tsx
@@ -1,7 +1,10 @@
 import React from 'react';
-import Main from './components/pages/Main/Main';
+import Main from './pages/Main/Main';
 import "./common/scss/main.scss"
 
-const App = (): JSX.Element => <Main />
+const App = (): JSX.Element =>
+    <>
+        <Main />
+    </>
 
 export default App;
