--- src/components/ui/button/Button.module.scss
@@ -1,17 +1,0 @@
-@import '../../../common/scss/abstracts/variables';
-
-.button {
-    color: $font-color;
-    border: none;
-    padding: 19px 31px;
-    border-radius: 50px;
-    text-transform: uppercase;
-    background-color: $button-color;
-    border: 2px solid $button-color;
-    font: $button-font;
-    cursor: pointer;
-    transition: .5s;
-    &:hover {
-        border-color: $font-color;
-    }
-}
