+++ src/pages/Main/Main.tsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import Welcome from './components/Welcome/Welcome';
 
-const Main = () => <Welcome />
+const Main = (): JSX.Element =>
+    <>
+        <Welcome />
+    </>
+
 
 export default Main;
