+++ src/pages/Main/components/Welcome/Welcome.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../common/scss/abstracts/variables';
+@import '../../../../common/scss/abstracts/variables';
 
 .header {
     font: $h1-font;
@@ -26,7 +26,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    background: url('../../../../../assets/images/Main/Welcome/main-background.png') no-repeat;
+    background: url('../../../../assets/images/Main/Welcome/main-background.png') no-repeat;
     background-position: bottom -40px right 40px;
 }
 
@@ -36,5 +36,5 @@
     margin-top: 86px;
     margin-bottom: 107px;
     width: 721px;
-    background: url('../../../../../assets/images/Main/Welcome/main.png');
+    background: url('../../../../assets/images/Main/Welcome/main.png');
 }
